Fix migration runner and flux_store_items schema
- migrate.ts: split SQL file by semicolons and execute each statement individually (conn.execute only runs one statement at a time)
- 006_flux.sql: change flux_store_items.value from INT to BIGINT to hold 5GB byte value (5368709120 exceeds INT max)

+++ backend/migrations/006_flux.sql
@@ -17,7 +17,7 @@ CREATE TABLE flux_store_items (
   description   TEXT         NULL,
   cost          DECIMAL(10,2) NOT NULL,
   type          ENUM('invite_token','freeleech_token','upload_credit','username_change') NOT NULL,
-  value         INT          DEFAULT 1,
+  value         BIGINT       DEFAULT 1,
   is_active     BOOLEAN      DEFAULT TRUE,
   display_order INT          DEFAULT 0,
   created_at    TIMESTAMP    DEFAULT CURRENT_TIMESTAMP
